+++ dask_geomodeling/tests/test_raster_spatial.py
@@ -79,6 +79,13 @@ def test_place_attrs(source, center):
     assert place.extent == pytest.approx(extent_epsg4326, rel=1e-4)
 
 
+def test_place_invalid_statistic(source, center):
+    with pytest.raises(ValueError):
+        place = raster.Place(
+            source, "EPSG:28992", center, [(50, 50)], statistic="nonexisting"
+        )
+
+
 def test_place_attrs_reproject(source, center_epsg3857):
     # Place should adapt the spatial attributes (extent & geometry)
     place = raster.Place(
@@ -147,15 +154,39 @@ def test_place_vertical_shift(source, center, vals_request):
     assert (values[:, 0, :] == 255).all()
 
 
-def test_place_multiple(source, center, vals_request):
+@pytest.mark.parametrize(
+    "statistic,expected",
+    [
+        ("first", (255, 7, 7, 7)),  # (no features, first, second, both)
+        ("last", (255, 7, 7, 7)),  # the default
+        ("count", (0, 1, 1, 2)),
+        ("sum", (0, 7, 7, 14)),
+        ("mean", (255, 7, 7, 7)),
+        ("min", (255, 7, 7, 7)),
+        ("max", (255, 7, 7, 7)),
+        ("argmin", (255, 0, 1, 0)),
+        ("argmax", (255, 0, 1, 0)),
+        ("std", (255, 0, 0, 0)),
+        ("var", (255, 0, 0, 0)),
+        ("median", (255, 7, 7, 7)),
+        ("p99", (255, 7, 7, 7)),
+    ],
+)
+def test_place_multiple(source, center, vals_request, statistic, expected):
     # place such that only the left and bottom ridges have values
-    place = raster.Place(source, "EPSG:28992", center, [(-40, 50), (50, -40)])
+    place = raster.Place(
+        source, "EPSG:28992", center, [(-40, 50), (50, -40)], statistic
+    )
     values = place.get_data(**vals_request)["values"]
     # swap Y axis for readable test
     values = values[:, ::-1, :]
-    assert (values[:, :10, 0] == 7).all()
-    assert (values[:, 0, :10] == 7).all()
-    assert (values[:, 1:, 1:] == 255).all()
+    # zero features
+    assert (values[:, 1:, 1:] == expected[0]).all()
+    # one feature
+    assert (values[:, 1:10, 0] == expected[1]).all()
+    assert (values[:, 0, 1:10] == expected[2]).all()
+    # two features
+    assert (values[:, 0, 0] == expected[3]).all()
 
 
 def test_place_outside(source, center, vals_request):
@@ -174,3 +205,44 @@ def test_place_time_request(source, center):
 def test_place_meta_request(source, center):
     place = raster.Place(source, "EPSG:28992", center, [(150, 50)])
     assert source.get_data(mode="meta") == place.get_data(mode="meta")
+
+
+@pytest.mark.parametrize("point,expected", [
+    ((5, 15), 7),  # zone 1
+    ((15, 15), 255),  # zone 2
+    ((5, 5), 255),  # zone 3
+    ((15, 5), 7),  # zone 4
+    ((10, 15), 255),  # line 1-2
+    ((5, 10), 255),  # line 1-3
+    ((15, 10), 7),  # line 2-4
+    ((10, 5), 7),  # line 3-4
+    ((10, 10), 7),  # center
+])
+def test_place_point_request(source, center, point, expected):
+    # For point requests, edges are important. Let's do a drawing:
+    # 20  _______
+    #    |       |
+    #    |   1   |   2
+    # 10 |_______|_______
+    #            |       |
+    #        3   |   4   |
+    #  0         |_______|
+    #    0       10     20
+    # - zone 1 and 4 are filled; zone 2 and 3 are empty (see below coordinates)
+    # A pixel includes its topleft corner and top and left edges
+    # - line between 2-4 and 3-4 are filled; line 1-2 and 1-3 are empty
+    # - center point at (10, 10) is filled
+    coordinates = [(60, -40), (-40, 60)]
+    place = raster.Place(
+        source, "EPSG:28992", anchor=center, coordinates=coordinates
+    )
+    point_request = dict(
+        mode="vals",
+        bbox=point * 2,
+        projection="EPSG:28992",
+        width=1,
+        height=1,
+    )
+    values = place.get_data(**point_request)["values"]
+    assert values.shape == (1, 1, 1)
+    assert values.item() == expected